+++ packages/code-infra/src/eslint/extensions.mjs
@@ -1,0 +1,5 @@
+export const EXTENSION_JS = '?(c|m)js?(x)';
+export const EXTENSION_TS = '?(c|m)[jt]s?(x)';
+export const EXTENSION_TS_ONLY = '?(c|m)ts?(x)';
+export const EXTENSION_DTS = `.d.${EXTENSION_TS_ONLY}`;
+export const EXTENSION_TEST_FILE = `.test.${EXTENSION_TS}`;
